requirements.txt
tensorflow
fastapi
uvicorn[standard]
pydantic
Pillow
python-multipart